fixes gemm source code (#1973)
* Add size_t support to gemm source code

* some more addition

+++ library/src/blas3/Tensile/gemm_source.hpp
@@ -101,15 +101,15 @@ namespace
                     {
                         if(TRANS_A == 'N')
                         {
-                            sA[n + thyA][m + thxA] = dA[i + j * lda];
+                            sA[n + thyA][m + thxA] = dA[i + j * size_t(lda)];
                         }
                         else if(TRANS_A == 'T')
                         {
-                            sA[n + thyA][m + thxA] = dA[i * lda + j];
+                            sA[n + thyA][m + thxA] = dA[i * size_t(lda) + j];
                         }
                         else if(TRANS_A == 'C')
                         {
-                            sA[n + thyA][m + thxA] = conj(dA[i * lda + j]);
+                            sA[n + thyA][m + thxA] = conj(dA[i * size_t(lda) + j]);
                         }
                     }
                     else
@@ -129,15 +129,15 @@ namespace
                     {
                         if(TRANS_B == 'N')
                         {
-                            sB[n + thyB][m + thxB] = dB[i + j * ldb];
+                            sB[n + thyB][m + thxB] = dB[i + j * size_t(ldb)];
                         }
                         else if(TRANS_B == 'T')
                         {
-                            sB[n + thyB][m + thxB] = dB[i * ldb + j];
+                            sB[n + thyB][m + thxB] = dB[i * size_t(ldb) + j];
                         }
                         else if(TRANS_B == 'C')
                         {
-                            sB[n + thyB][m + thxB] = conj(dB[i * ldb + j]);
+                            sB[n + thyB][m + thxB] = conj(dB[i * size_t(ldb) + j]);
                         }
                     }
                     else
@@ -167,12 +167,12 @@ namespace
                 {
                     if(BETA_EQ_ZERO)
                     {
-                        dC[coord_dCn * ldc + coord_dCm] = alpha * rC[n][m];
+                        dC[coord_dCn * size_t(ldc) + coord_dCm] = alpha * rC[n][m];
                     }
                     else
                     {
-                        dC[coord_dCn * ldc + coord_dCm]
-                            = alpha * rC[n][m] + beta * dC[coord_dCn * ldc + coord_dCm];
+                        dC[coord_dCn * size_t(ldc) + coord_dCm]
+                            = alpha * rC[n][m] + beta * dC[coord_dCn * size_t(ldc) + coord_dCm];
                     }
                 }
             }
@@ -238,14 +238,14 @@ namespace
 
         size_t coord_A, coord_B;
         if(TRANS_A == 'N')
-            coord_A = (thxA + blx * BLK_M) + (thyA)*lda;
+            coord_A = (thxA + blx * BLK_M) + (thyA)*size_t(lda);
         else if(TRANS_A == 'T' || TRANS_A == 'C')
-            coord_A = (thxA + blx * BLK_M) * lda + (thyA);
+            coord_A = (thxA + blx * BLK_M) * size_t(lda) + (thyA);
 
         if(TRANS_B == 'N')
-            coord_B = thxB + (bly * BLK_N + thyB) * ldb;
+            coord_B = thxB + (bly * BLK_N + thyB) * size_t(ldb);
         else if(TRANS_B == 'T' || TRANS_B == 'C')
-            coord_B = thxB * ldb + (bly * BLK_N + thyB);
+            coord_B = thxB * size_t(ldb) + (bly * BLK_N + thyB);
 
         int kk = 0;
         for(; kk < K; kk += BLK_K)
@@ -254,30 +254,30 @@ namespace
                 for(int m = 0; m < BLK_M; m += DIM_M_A)
                     if(TRANS_A == 'N')
                     {
-                        sA[n + thyA][m + thxA] = dA[coord_A + m + n * lda];
+                        sA[n + thyA][m + thxA] = dA[coord_A + m + n * size_t(lda)];
                     }
                     else if(TRANS_A == 'T')
                     {
-                        sA[n + thyA][m + thxA] = dA[coord_A + m * lda + n];
+                        sA[n + thyA][m + thxA] = dA[coord_A + m * size_t(lda) + n];
                     }
                     else if(TRANS_A == 'C')
                     {
-                        sA[n + thyA][m + thxA] = conj(dA[coord_A + m * lda + n]);
+                        sA[n + thyA][m + thxA] = conj(dA[coord_A + m * size_t(lda) + n]);
                     }
 
             for(int n = 0; n < BLK_N; n += DIM_N_B)
                 for(int m = 0; m < BLK_K; m += DIM_M_B)
                     if(TRANS_B == 'N')
                     {
-                        sB[n + thyB][m + thxB] = dB[coord_B + m + n * ldb];
+                        sB[n + thyB][m + thxB] = dB[coord_B + m + n * size_t(ldb)];
                     }
                     else if(TRANS_B == 'T')
                     {
-                        sB[n + thyB][m + thxB] = dB[coord_B + m * ldb + n];
+                        sB[n + thyB][m + thxB] = dB[coord_B + m * size_t(ldb) + n];
                     }
                     else if(TRANS_B == 'C')
                     {
-                        sB[n + thyB][m + thxB] = conj(dB[coord_B + m * ldb + n]);
+                        sB[n + thyB][m + thxB] = conj(dB[coord_B + m * size_t(ldb) + n]);
                     }
 
             __syncthreads();
@@ -290,14 +290,14 @@ namespace
             __syncthreads();
 
             if(TRANS_A == 'N')
-                coord_A += BLK_K * lda;
+                coord_A += BLK_K * size_t(lda);
             else if(TRANS_A == 'T' || TRANS_A == 'C')
                 coord_A += BLK_K;
 
             if(TRANS_B == 'N')
                 coord_B += BLK_K;
             else if(TRANS_B == 'T' || TRANS_B == 'C')
-                coord_B += BLK_K * ldb;
+                coord_B += BLK_K * size_t(ldb);
         }
 
         for(int n = 0; n < BLK_N / DIM_N; ++n)
@@ -309,12 +309,12 @@ namespace
 
                 if(BETA_EQ_ZERO)
                 {
-                    dC[coord_dCn * ldc + coord_dCm] = alpha * rC[n][m];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm] = alpha * rC[n][m];
                 }
                 else
                 {
-                    dC[coord_dCn * ldc + coord_dCm]
-                        = alpha * rC[n][m] + beta * dC[coord_dCn * ldc + coord_dCm];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm]
+                        = alpha * rC[n][m] + beta * dC[coord_dCn * size_t(ldc) + coord_dCm];
                 }
             }
         }
@@ -374,13 +374,13 @@ namespace
 
         size_t coord_A, coord_B;
         if(TRANS_A == 'N')
-            coord_A = (blx * BLK_M + thxA) + thyA * lda;
+            coord_A = (blx * BLK_M + thxA) + thyA * size_t(lda);
         else if(TRANS_A == 'T' || TRANS_A == 'C')
-            coord_A = (blx * BLK_M + thxA) * lda + thyA;
+            coord_A = (blx * BLK_M + thxA) * size_t(lda) + thyA;
         if(TRANS_B == 'N')
-            coord_B = (bly * BLK_N + thyB) * ldb + thxB;
+            coord_B = (bly * BLK_N + thyB) * size_t(ldb) + thxB;
         else if(TRANS_B == 'T' || TRANS_B == 'C')
-            coord_B = (bly * BLK_N + thyB) + thxB * ldb;
+            coord_B = (bly * BLK_N + thyB) + thxB * size_t(ldb);
 
         for(int n = 0; n < BLK_N / DIM_N; ++n)
             for(int m = 0; m < BLK_M / DIM_M; ++m)
@@ -392,20 +392,20 @@ namespace
             for(int n = 0; n < BLK_K; n += DIM_N_A)
                 for(int m = 0; m < BLK_M; m += DIM_M_A)
                     if(TRANS_A == 'N')
-                        sA[n + thyA][m + thxA] = dA[coord_A + (n * lda + m)];
+                        sA[n + thyA][m + thxA] = dA[coord_A + (n * size_t(lda) + m)];
                     else if(TRANS_A == 'T')
-                        sA[n + thyA][m + thxA] = dA[coord_A + (n + m * lda)];
+                        sA[n + thyA][m + thxA] = dA[coord_A + (n + m * size_t(lda))];
                     else if(TRANS_A == 'C')
-                        sA[n + thyA][m + thxA] = conj(dA[coord_A + (n + m * lda)]);
+                        sA[n + thyA][m + thxA] = conj(dA[coord_A + (n + m * size_t(lda))]);
 
             for(int n = 0; n < BLK_N; n += DIM_N_B)
                 for(int m = 0; m < BLK_K; m += DIM_M_B)
                     if(TRANS_B == 'N')
-                        sB[n + thyB][m + thxB] = dB[coord_B + (n * ldb + m)];
+                        sB[n + thyB][m + thxB] = dB[coord_B + (n * size_t(ldb) + m)];
                     else if(TRANS_B == 'T')
-                        sB[n + thyB][m + thxB] = dB[coord_B + (n + m * ldb)];
+                        sB[n + thyB][m + thxB] = dB[coord_B + (n + m * size_t(ldb))];
                     else if(TRANS_B == 'C')
-                        sB[n + thyB][m + thxB] = conj(dB[coord_B + (n + m * ldb)]);
+                        sB[n + thyB][m + thxB] = conj(dB[coord_B + (n + m * size_t(ldb))]);
 
             __syncthreads();
 
@@ -417,14 +417,14 @@ namespace
             __syncthreads();
 
             if(TRANS_A == 'N')
-                coord_A += BLK_K * lda;
+                coord_A += BLK_K * size_t(lda);
             else if(TRANS_A == 'T' || TRANS_A == 'C')
                 coord_A += BLK_K;
 
             if(TRANS_B == 'N')
                 coord_B += BLK_K;
             else if(TRANS_B == 'T' || TRANS_B == 'C')
-                coord_B += BLK_K * ldb;
+                coord_B += BLK_K * size_t(ldb);
         }
 
         for(int n = 0; n < BLK_N / DIM_N; ++n)
@@ -436,19 +436,20 @@ namespace
 
                 if(alpha == 1 && beta == 1)
                 {
-                    dC[coord_dCn * ldc + coord_dCm] += rC[n][m];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm] += rC[n][m];
                 }
                 else if(alpha == 1 && beta == -1)
                 {
-                    dC[coord_dCn * ldc + coord_dCm] = -dC[coord_dCn * ldc + coord_dCm] + rC[n][m];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm]
+                        = -dC[coord_dCn * size_t(ldc) + coord_dCm] + rC[n][m];
                 }
                 else if(alpha == -1 && beta == 0)
                 {
-                    dC[coord_dCn * ldc + coord_dCm] = -rC[n][m];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm] = -rC[n][m];
                 }
                 else if(alpha == 1 && beta == 0)
                 {
-                    dC[coord_dCn * ldc + coord_dCm] = rC[n][m];
+                    dC[coord_dCn * size_t(ldc) + coord_dCm] = rC[n][m];
                 }
             }
         }
